fix memory leak shutting down the channel after the process exited
This also rearranges the channel API functions to all live in channel.c,
removes a few unused callbacks and methods.
