+++ mettle/src/coreapi.c
@@ -4,7 +4,6 @@
  * @file tlv_coreapi.c
  */
 
-#include "channel.h"
 #include "log.h"
 #include "tlv.h"
 
@@ -50,221 +49,6 @@ static struct tlv_packet *core_shutdown(struct tlv_handler_ctx *ctx)
 	return p;
 }
 
-static struct tlv_packet *core_channel_open(struct tlv_handler_ctx *ctx)
-{
-	struct mettle *m = ctx->arg;
-	struct channelmgr *cm = mettle_get_channelmgr(m);
-
-	char *channel_type = tlv_packet_get_str(ctx->req, TLV_TYPE_CHANNEL_TYPE);
-	if (channel_type == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel *c = channelmgr_channel_new(cm, channel_type);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-	ctx->channel = c;
-	ctx->channel_id = channel_get_id(c);
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	struct tlv_packet *p;
-	if (cbs->new_cb && cbs->new_cb(ctx, c) == -1) {
-		channel_free(c);
-		p = tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	} else {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-	}
-	return p;
-}
-
-static struct channel * get_channel_by_id(struct tlv_handler_ctx *ctx)
-{
-	struct mettle *m = ctx->arg;
-	struct channelmgr *cm = mettle_get_channelmgr(m);
-
-	if (tlv_packet_get_u32(ctx->req, TLV_TYPE_CHANNEL_ID, &ctx->channel_id)) {
-		return NULL;
-	}
-	ctx->channel = channelmgr_channel_by_id(cm, ctx->channel_id);
-	return ctx->channel;
-}
-
-static struct tlv_packet *core_channel_close(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	struct tlv_packet *p;
-	if (cbs->free_cb && cbs->free_cb(c) == -1) {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	} else {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-	}
-	channel_free(c);
-	return p;
-}
-
-static struct tlv_packet *core_channel_interact(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-	bool interact = false;
-	tlv_packet_get_bool(ctx->req, TLV_TYPE_BOOL, &interact);
-
-	channel_set_interactive(c, interact);
-
-	return tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-}
-
-static struct tlv_packet *core_channel_eof(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	struct tlv_packet *p;
-	if (cbs->eof_cb) {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-		p = tlv_packet_add_bool(p, TLV_TYPE_BOOL, cbs->eof_cb(c));
-	} else {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-	return p;
-}
-
-static struct tlv_packet *core_channel_seek(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	uint32_t offset, whence;
-	if (tlv_packet_get_u32(ctx->req, TLV_TYPE_SEEK_OFFSET, &offset) == -1 ||
-	    tlv_packet_get_u32(ctx->req, TLV_TYPE_SEEK_WHENCE, &whence) == -1) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_EINVAL);
-	}
-
-	struct tlv_packet *p;
-	if (cbs->seek_cb) {
-		int rc = cbs->seek_cb(c, offset, whence);
-		p = tlv_packet_response_result(ctx,
-			rc == 0 ? TLV_RESULT_SUCCESS : TLV_RESULT_FAILURE);
-	} else {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-	return p;
-}
-
-static struct tlv_packet *core_channel_tell(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	struct tlv_packet *p;
-	if (cbs->tell_cb) {
-		ssize_t offset = cbs->tell_cb(c);
-		if (offset >= 0) {
-			p = tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-			p = tlv_packet_add_u32(p, TLV_TYPE_SEEK_POS, offset);
-		} else {
-			p = tlv_packet_response_result(ctx, errno);
-		}
-	} else {
-		p = tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-	return p;
-}
-
-static struct tlv_packet *core_channel_read(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	uint32_t len = 0;
-	if (tlv_packet_get_u32(ctx->req, TLV_TYPE_LENGTH, &len) == -1) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	if (cbs->read_cb) {
-		char *buf = calloc(1, len);
-		if (buf == NULL) {
-			return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-		}
-
-		ssize_t bytes_read = cbs->read_cb(c, buf, len);
-		struct tlv_packet *p;
-		if (bytes_read >= 0) {
-			p = tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-			p = tlv_packet_add_raw(p, TLV_TYPE_CHANNEL_DATA, buf, bytes_read);
-		} else {
-			p = tlv_packet_response_result(ctx, errno);
-		}
-		free(buf);
-		return p;
-	} else {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-}
-
-static struct tlv_packet *core_channel_write(struct tlv_handler_ctx *ctx)
-{
-	struct channel *c = get_channel_by_id(ctx);
-	if (c == NULL) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	uint32_t len = 0;
-	if (tlv_packet_get_u32(ctx->req, TLV_TYPE_LENGTH, &len) == -1) {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-
-	struct channel_callbacks *cbs = channel_get_callbacks(c);
-
-	if (cbs->write_cb) {
-		size_t buf_len = 0;
-		char *buf = tlv_packet_get_raw(ctx->req, TLV_TYPE_CHANNEL_DATA, &buf_len);
-		if (buf == NULL) {
-			return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-		}
-
-		ssize_t bytes_written = cbs->write_cb(c, buf, len);
-		struct tlv_packet *p;
-		if (bytes_written > 0) {
-			p = tlv_packet_response_result(ctx, TLV_RESULT_SUCCESS);
-			p = tlv_packet_add_u32(p, TLV_TYPE_LENGTH, bytes_written);
-		} else {
-			p = tlv_packet_response_result(ctx, errno);
-		}
-		return p;
-	} else {
-		return tlv_packet_response_result(ctx, TLV_RESULT_FAILURE);
-	}
-}
-
 static struct tlv_packet *core_machine_id(struct tlv_handler_ctx *ctx)
 {
 	struct mettle *m = ctx->arg;
@@ -296,12 +80,4 @@ void tlv_register_coreapi(struct mettle *m)
 	tlv_dispatcher_add_handler(td, "core_machine_id", core_machine_id, m);
 	tlv_dispatcher_add_handler(td, "core_uuid", core_uuid, m);
 	tlv_dispatcher_add_handler(td, "core_shutdown", core_shutdown, m);
-	tlv_dispatcher_add_handler(td, "core_channel_open", core_channel_open, m);
-	tlv_dispatcher_add_handler(td, "core_channel_eof", core_channel_eof, m);
-	tlv_dispatcher_add_handler(td, "core_channel_seek", core_channel_seek, m);
-	tlv_dispatcher_add_handler(td, "core_channel_tell", core_channel_tell, m);
-	tlv_dispatcher_add_handler(td, "core_channel_read", core_channel_read, m);
-	tlv_dispatcher_add_handler(td, "core_channel_write", core_channel_write, m);
-	tlv_dispatcher_add_handler(td, "core_channel_close", core_channel_close, m);
-	tlv_dispatcher_add_handler(td, "core_channel_interact", core_channel_interact, m);
 }
